+++ core/src/com/mygdx/game/Controller/Entitys/DinamicObjects/PressingPlateBody.java
@@ -19,7 +19,7 @@ public class PressingPlateBody {
     private Body body;
     private BodyDef bdef;
     private FixtureDef fdef;
-    public PressingPlateBody(World world, PressingPlate pressingPlate, int x, int y) {
+    public PressingPlateBody(World world, PressingPlate pressingPlate, float x, float y) {
         bdef = new BodyDef();
         bdef.position.set(x * MyGame.PIXEL_TO_METER, y * MyGame.PIXEL_TO_METER);
         bdef.type = BodyDef.BodyType.StaticBody;
